+++ src/components/Acciones/Articulo2/MainSectionArticulo2.jsx
@@ -1,49 +1,44 @@
 import Image from "next/image";
 
-export default function MainSectionArticulo2({ image1, image2 }) {
+export default function MainSectionArticulo1({ image }) {
   return (
     <>
-      <article>
-        <h1 className=" text-green  lg:text-2xl md:text-xl text-lg">
-          FERIA DE VILLA DOMINICO
-        </h1>
-      </article>
-      <article>
-        <p>
-          El día 21 de diciembre de 2014 sería el comienzo de la caída de uno de
-          los máximos exponentes en la venta de aves: la Feria De Villa
-          Domínico, situada en la zona sur del conurbano bonaerense, lugar de
-          asidua concurrencia de las personas que no tienen reparos en comprar y
-          vender ejemplares de aves víctimas del tráfico de fauna, actividad
-          prohibida por leyes y ordenanzas, y repudiada por gran parte de la
-          comunidad. Desde bastante tiempo atrás se venía trabajando en la
-          búsqueda de una manera eficiente de terminar con uno de los focos más
-          grandes de venta delictiva de animales. Un grupo de vecinos de Villa
-          Domínico, las ONG “APAC Avellaneda” y “Zoonosis Lanús”, persistieron
-          en las denuncias hasta dar con una Fiscal responsable y sensible
-          frente al maltrato animal, la Dra. Carla Musitani. Su forma de
-          intervención permitió que se pudiera embestir de manera persistente
-          con allanamientos que se llevaron a cabo entre diciembre de 2014 y
-          abril de 2015, con una acción judicial por maltrato animal
-          (incumplimiento de la Ley 14.346). La Asociación Civil de Ayuda a las
-          Aves Pájaros Caídos fue convocada para salvaguardar las aves que allí
-          se encontraran, por lo que concurrimos con nuestro equipo de rescate,
-          acompañando los allanamientos. Colaboramos en el traslado e inventario
-          de las aves, asistiéndolas de manera tal de evitarles en lo posible,
-          el estrés que estas situaciones producen. Jilgueros, cabecita negra,
-          reyes del bosque, cardenales, cardenillas, reinamoras, tordo amarillo,
-          federales, tordos músicos, etc. hacinados en jaulas pequeñísimas,
-          sucias, al rayo del sol o en los baúles de los autos, sin agua ni
-          comida adecuada, algunos con heridas, fue el panorama tenebroso con el
-          que nos encontramos en el deambular por los puestos de la feria.
+      <article className=" px-4 w-11/12 md:p-0 md:w-6/12">
+        <p className=" break-words">
+          `25-07-2023 INCAUTACION DE GUACAMAYOS EN MATADEROS Efectivos del
+          Departamento Delitos Ambientales de la Policía Federal Argentina,
+          rescataron dos Guacamayos Azul y Amarillo (Ara Ararauna) en peligro de
+          extinción, los cuales se encontraban cautivos en jaulas escondidas
+          entre los acopios de mercadería de una fábrica de vidrios de la
+          localidad de Mataderos, CABA, en infracción a las Leyes de
+          Conservación de la Fauna Silvestre (22.421) y la de Maltrato Animal
+          (14.346) La investigación comenzó a raíz de una denuncia en la que se
+          hizo saber a las autoridades de la existencia de estos dos Guacamayos
+          enjaulados en un área no adecuada. Tomó intervención la Unidad Fiscal
+          de Investigación en Materia Ambiental (UFEMA) de la Ciudad Autónoma de
+          Buenos Aires a cargo del Dr. Carlos Fel Rolero Santurian, Secretaría
+          del Dr. Ricardo Bomparola, quien dio intervención al Departamento de
+          Delitos Ambientales, lográndose determinar la veracidad de la
+          denuncia. Con las pruebas reunidas, se procedió a la incautación de
+          las dos aves, que fueron trasladadas a la Ong. Pájaros Caídos, en
+          donde se procurará un período de cuarentena para su rehabilitación y
+          posterior reinserción a su hábitat especifico. El responsable de la
+          firma quedó a disposición de la Fiscalía interventora, afectado por
+          los delitos de Tráfico de Fauna Silvestre y Maltrato Animal. La manera
+          de frenar el Tráfico Ilegal de Fauna es hacer la Denuncia!!! Si se
+          denuncia, las autoridades pueden intervenir y salvar a los animales
+          que se tienen en cautiverio. Si ves maltrato animal, venta de animales
+          silvestres, te enterás de caza ilegal de fauna silvestre, por favor
+          Denunciá, las denuncias si sirven.`
         </p>
       </article>
-      <article className="flex justify-around">
-        <Image src={image1} className="w-5/12 h-auto" alt="img1" />
+
+      <article className="w-6/12 hidden md:block">
         <Image
-          src={image2}
-          className="w-5/12 h-auto hidden md:block"
-          alt="img2"
+          src={image}
+          alt="Bg-article-1"
+          className="m-auto w-8/12 h-auto max-h-[625px]"
+          quality={100}
         />
       </article>
     </>
